move followbutton to button
